# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Novel Dialogue Reader API",
    "description": "轻小说对话辅助阅读器本地后端。契约摘要见 docs/CONTRACTS.md，完整规格见 DEVELOPMENT.md。",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/jobs": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "创建任务",
        "operationId": "create_job_route_api_jobs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataEnvelope_JobDetailOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{job_id}": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "任务状态与用量",
        "operationId": "get_job_route_api_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataEnvelope_JobDetailOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{job_id}/pause": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Pause Job Route",
        "operationId": "pause_job_route_api_jobs__job_id__pause_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataEnvelope_JobDetailOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{job_id}/resume": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Resume Job Route",
        "operationId": "resume_job_route_api_jobs__job_id__resume_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataEnvelope_JobRunOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{job_id}/run": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "立即执行（测试/手动）",
        "operationId": "run_job_route_api_jobs__job_id__run_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataEnvelope_JobRunOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{job_id}/reconcile": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Reconcile Job Route",
        "operationId": "reconcile_job_route_api_jobs__job_id__reconcile_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReconcileIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataEnvelope_dict_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{job_id}/recovery": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "非完成状态的恢复动作（含是否付费与是否缺凭据）",
        "description": "把任务状态翻译成可执行动作：读接口，不调用模型、不改任务。",
        "operationId": "job_recovery_route_api_jobs__job_id__recovery_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataEnvelope_JobRecoveryOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/books/{book_id}/estimates": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "本地估算（不调用模型）",
        "operationId": "estimate_route_api_books__book_id__estimates_post",
        "parameters": [
          {
            "name": "book_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Book Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EstimateIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataEnvelope_EstimateOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/books/{book_id}/usage": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "按任务汇总用量",
        "operationId": "usage_route_api_books__book_id__usage_get",
        "parameters": [
          {
            "name": "book_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Book Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataEnvelope_UsageOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BudgetIn": {
        "properties": {
          "max_input_tokens": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Input Tokens"
          },
          "max_output_tokens": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Output Tokens"
          },
          "max_rechecks": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Max Rechecks",
            "default": 0
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "BudgetIn"
      },
...
      "DataEnvelope_EstimateOut_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/EstimateOut"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "description": "本次请求的追踪 ID，与响应头 X-Request-ID 一致。"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "data",
          "request_id"
        ],
        "title": "DataEnvelope[EstimateOut]"
      },
...
      "DataEnvelope_JobDetailOut_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/JobDetailOut"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "description": "本次请求的追踪 ID，与响应头 X-Request-ID 一致。"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "data",
          "request_id"
        ],
        "title": "DataEnvelope[JobDetailOut]"
      },
      "DataEnvelope_JobRecoveryOut_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/JobRecoveryOut"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "description": "本次请求的追踪 ID，与响应头 X-Request-ID 一致。"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "data",
          "request_id"
        ],
        "title": "DataEnvelope[JobRecoveryOut]"
      },
      "DataEnvelope_JobRunOut_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/JobRunOut"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "description": "本次请求的追踪 ID，与响应头 X-Request-ID 一致。"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "data",
          "request_id"
        ],
        "title": "DataEnvelope[JobRunOut]"
      },
...
      "DataEnvelope_UsageOut_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/UsageOut"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "description": "本次请求的追踪 ID，与响应头 X-Request-ID 一致。"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "data",
          "request_id"
        ],
        "title": "DataEnvelope[UsageOut]"
      },
      "DataEnvelope_dict_": {
        "properties": {
          "data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Data"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "description": "本次请求的追踪 ID，与响应头 X-Request-ID 一致。"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "data",
          "request_id"
        ],
        "title": "DataEnvelope[dict]"
      },
...
      "EstimateIn": {
        "properties": {
          "book_version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Book Version Id"
          },
          "range": {
            "additionalProperties": true,
            "type": "object",
            "title": "Range"
          },
          "reading_mode": {
            "$ref": "#/components/schemas/ReadingMode",
            "default": "initial"
          },
          "visible_horizon_cp": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Visible Horizon Cp"
          },
          "budget": {
            "$ref": "#/components/schemas/BudgetIn"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "EstimateIn"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobCreate": {
        "properties": {
          "book_id": {
            "type": "string",
            "title": "Book Id"
          },
          "kind": {
            "$ref": "#/components/schemas/JobKind",
            "default": "INFERENCE"
          },
          "mode": {
            "$ref": "#/components/schemas/JobPurpose",
            "description": "preview / process",
            "default": "process"
          },
          "book_version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Book Version Id"
          },
          "range": {
            "additionalProperties": true,
            "type": "object",
            "title": "Range",
            "description": "{start_cp, end_cp, chapter_id?}"
          },
          "profile_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Id"
          },
          "reading_mode": {
            "$ref": "#/components/schemas/ReadingMode",
            "default": "initial"
          },
          "visible_horizon_cp": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Visible Horizon Cp"
          },
          "budget": {
            "$ref": "#/components/schemas/BudgetIn"
          },
          "idempotency_key": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Idempotency Key"
          },
          "run_now": {
            "type": "boolean",
            "title": "Run Now",
            "description": "是否立即在后台执行（测试/E2E 可显式触发）",
            "default": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "book_id",
          "idempotency_key"
        ],
        "title": "JobCreate"
      },
      "JobDetailOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "kind": {
            "$ref": "#/components/schemas/JobKind"
          },
          "purpose": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobPurpose"
              },
              {
                "type": "null"
              }
            ]
          },
          "state": {
            "$ref": "#/components/schemas/JobState"
          },
          "book_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Book Id"
          },
          "book_version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Book Version Id"
          },
          "progress": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Progress"
          },
          "checkpoint": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checkpoint"
          },
          "last_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error"
          },
          "windows": {
            "items": {
              "$ref": "#/components/schemas/JobWindowOut"
            },
            "type": "array",
            "title": "Windows"
          },
          "remaining_windows": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Remaining Windows",
            "default": 0
          },
          "windows_total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Windows Total",
            "default": 0
          },
          "calls": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Calls",
            "default": 0
          },
          "cached_windows": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Cached Windows",
            "default": 0
          },
          "unknown_usage_runs": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Unknown Usage Runs",
            "default": 0
          },
          "usage": {
            "additionalProperties": true,
            "type": "object",
            "title": "Usage"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "kind",
          "state",
          "created_at",
          "updated_at"
        ],
        "title": "JobDetailOut"
      },
      "JobKind": {
        "type": "string",
        "enum": [
          "IMPORT",
          "INFERENCE",
          "CHARACTER_ROSTER",
          "RECHECK",
          "RECOMPUTE",
          "EXPORT"
        ],
        "title": "JobKind"
      },
      "JobPurpose": {
        "type": "string",
        "enum": [
          "preview",
          "process",
          "none"
        ],
        "title": "JobPurpose",
        "description": "INFERENCE 的目的；preview/process 共用同一识别引擎与缓存。"
      },
      "JobRecoveryOut": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "state": {
            "$ref": "#/components/schemas/JobState"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "actions": {
            "items": {
              "$ref": "#/components/schemas/RecoveryActionOut"
            },
            "type": "array",
            "title": "Actions"
          },
          "windows_total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Windows Total",
            "default": 0
          },
          "windows_done": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Windows Done",
            "default": 0
          },
          "remaining_windows": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Remaining Windows",
            "default": 0
          },
          "unknown_runs": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Unknown Runs",
            "default": 0
          },
          "unknown_usage_runs": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Unknown Usage Runs",
            "default": 0
          },
          "retry_in_seconds": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Retry In Seconds",
            "description": "限流退避的建议等待时间（0 表示可立即继续）"
          },
          "requires_credential": {
            "type": "boolean",
            "title": "Requires Credential",
            "description": "是否因为缺少模型凭据而失败（需要先去补充密钥）",
            "default": false
          },
          "last_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "job_id",
          "state",
          "summary",
          "updated_at"
        ],
        "title": "JobRecoveryOut"
      },
      "JobRunOut": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "state": {
            "$ref": "#/components/schemas/JobState"
          },
          "windows_total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Windows Total",
            "default": 0
          },
          "windows_done": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Windows Done",
            "default": 0
          },
          "cached_windows": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Cached Windows",
            "default": 0
          },
          "calls": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Calls",
            "default": 0
          },
          "unknown_runs": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Unknown Runs",
            "default": 0
          },
          "budget_exhausted": {
            "type": "boolean",
            "title": "Budget Exhausted",
            "default": false
          },
          "errors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Errors"
          },
          "usage": {
            "additionalProperties": true,
            "type": "object",
            "title": "Usage"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "job_id",
          "state"
        ],
        "title": "JobRunOut"
      },
      "JobState": {
        "type": "string",
        "enum": [
          "QUEUED",
          "RUNNING",
          "PAUSING",
          "PAUSED",
          "PARTIAL",
          "COMPLETED",
          "FAILED",
          "BUDGET_EXHAUSTED",
          "NEEDS_RECONCILIATION"
        ],
        "title": "JobState"
      },
...
      "ReadingMode": {
        "type": "string",
        "enum": [
          "initial",
          "reread"
        ],
        "title": "ReadingMode"
      },
...
      "ReconcileIn": {
        "properties": {
          "action": {
            "type": "string",
            "enum": [
              "retry",
              "keep_unknown"
            ],
            "title": "Action",
            "default": "keep_unknown"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ReconcileIn"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}